third_party/WebKit/LayoutTests/http/tests/pointer-lock/iframe-sandboxed-expected.txt

CONSOLE ERROR: Blocked pointer lock on an element because the element's frame is sandboxed and the 'allow-pointer-lock' permission is not set.
Test sandboxed iframe blocks pointer lock.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS message is "inner-iframe.html onpointerlockerror"
PASS successfullyParsed is true

TEST COMPLETE
